+++ entrypoints/SettingComponents/BasicSettings/MenuBookmark.vue
@@ -53,14 +53,14 @@ export default {
     if (this.modelValue) {
       const vm = this;
       this.bookmarkIntervalId = setInterval(() => {
-        if ($(".linxudoscripts-bookmark").length < 1) {
+        if ($(".linuxdoscripts-bookmark").length < 1) {
           const btn = $(
-            `<button class="btn btn-icon-text linxudoscripts-bookmark" type="button"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"  fill="none"  stroke="currentColor" stroke-width="2" stroke-linecap="round"  stroke-linejoin="round" class="d-icon icon icon-tabler icons-tabler-outline icon-tabler-bookmark-plus"><path stroke="none" d="M0 0h24v24H0z" fill="none"/><path d="M12 17l-6 4v-14a4 4 0 0 1 4 -4h4a4 4 0 0 1 4 4v5" /><path d="M16 19h6" /><path d="M19 16v6" /></svg><span class="d-button-label">收藏</span></button>`
+            `<button class="btn btn-icon-text linuxdoscripts-bookmark" type="button"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"  fill="none"  stroke="currentColor" stroke-width="2" stroke-linecap="round"  stroke-linejoin="round" class="d-icon icon icon-tabler icons-tabler-outline icon-tabler-bookmark-plus"><path stroke="none" d="M0 0h24v24H0z" fill="none"/><path d="M12 17l-6 4v-14a4 4 0 0 1 4 -4h4a4 4 0 0 1 4 4v5" /><path d="M16 19h6" /><path d="M19 16v6" /></svg><span class="d-button-label">收藏</span></button>`
           );
 
-          $(".linxudoscripts-btn").append(btn);
+          $(".linuxdoscripts-btn").append(btn);
 
-          $(".linxudoscripts-bookmark").click(function () {
+          $(".linuxdoscripts-bookmark").click(function () {
             if ($(".header-title .topic-link").text().trim() == "") {
               vm.messageToast("请将页面往下滑动一点点，再点击收藏！");
               return false;